Move logic to ec::encoding::sec1 module
