+++ aws-lc-rs/src/cbb.rs
@@ -1,9 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR ISC
 
+use crate::aws_lc::{CBB_cleanup, CBB_finish, CBB_init, CBB};
 use crate::error::Unspecified;
 use crate::ptr::LcPtr;
-use crate::aws_lc::{CBB_cleanup, CBB_finish, CBB_init, CBB};
 use core::marker::PhantomData;
 use core::mem::MaybeUninit;
 use core::ptr::null_mut;
